Fix an issue where the transaction number was being cast from a long to an int.  VPay transaction numbers have crossed the 32-bit threshold and 32 bit values won't hold the data anymore

+++ VPay.Payment.Common/Models/HeaderData.cs
@@ -2,7 +2,7 @@
 {
     public class HeaderData
     {
-        public int TransNumber { get; set; }
+        public long TransNumber { get; set; }
         public string Client { get; set; }
         public string BillCode { get; set; }
         public string BillType { get; set; }
